Contribution from Georg Fischer, rowsums.2021-11-09.23.tgz

+++ src/irvine/oeis/a059/A059036.java
@@ -15,9 +15,9 @@ public class A059036 extends UpperLeftTriangle {
   public A059036() {
     super(0, 0, -1);
   }
-
+  
   @Override
   public Z matrixElement(final int n, final int k) {
-    return Z.valueOf(n + k + n * k);
+    return Z.valueOf(n + k + n * (long) k);
   }
 }
